✨ Feature: login -> auth 폴더명 변경 및 auth 쿼리 key 생성
