+++ src/features/auth/apis.ts
@@ -1,0 +1,11 @@
+import api from '@/axios/instance';
+import type { User } from '@features/user/types';
+
+const authApis = {
+  verifyAuth: async (): Promise<User> => {
+    const response = await api.get('/auth/verify');
+    return response.data;
+  },
+};
+
+export default authApis;
