+++ src/features/auth/queries.ts
@@ -1,0 +1,17 @@
+import { useQuery } from '@tanstack/react-query';
+import authApis from './apis';
+
+const authKeys = {
+  verify: ['auth', 'verify'] as const,
+};
+
+export const authQuery = {
+  verify: () => {
+    return useQuery({
+      queryKey: authKeys.verify,
+      queryFn: authApis.verifyAuth,
+      staleTime: 0,
+      gcTime: 0,
+    });
+  },
+};
