+++ src/features/auth/service/handleLogout.ts
