+++ src/pages/quiz/Quiz.tsx
@@ -19,8 +19,8 @@ import { quizzesQuery } from '@features/quiz/queries';
 import { userQuizzesQuery } from '@features/user/queries';
 import Header from '@common/layout/Header';
 import ProgressBar from '@features/progress/ui/ProgressBar';
-import LoginPrompt from '@features/login/ui/LoginPrompt';
-import Login from '@features/login/ui/Login';
+import LoginPrompt from '@/features/auth/ui/LoginPrompt';
+import Login from '@/features/auth/ui/Login';
 import Question from '@features/quiz/ui/Question';
 import Combination from '@features/quiz/ui/Combination';
 import MultipleChoice from '@features/quiz/ui/MultipleChoice';
